+++ src/class_diagram/mod.rs
@@ -3,7 +3,10 @@ use crate::checker::Checker;
 use crate::class_helpers::{ClassDefHelpers, QualifiedNameHelpers};
 use crate::class_type_detector::ClassTypeDetector;
 use crate::parameter_generator::ParameterGenerator;
-use crate::renderer::*;
+use crate::renderer::{
+    Attribute, ClassNode, CompositionEdge, Diagram, DiagramDirection, MethodSignature,
+    RelationType, RelationshipEdge, Visibility,
+};
 use crate::type_analyzer;
 use indexmap::IndexSet;
 use ruff_linter::source_kind::SourceKind;
@@ -37,19 +40,21 @@ enum BaseKind {
 
 pub struct ClassDiagram {
     diagram: Diagram,
+    direction: DiagramDirection,
     pub path: String,
 }
 
 impl Default for ClassDiagram {
     fn default() -> Self {
-        Self::new()
+        Self::new(DiagramDirection::default())
     }
 }
 
 impl ClassDiagram {
-    pub fn new() -> Self {
+    pub fn new(direction: DiagramDirection) -> Self {
         Self {
-            diagram: Diagram::new(None),
+            diagram: Diagram::new(),
+            direction,
             path: String::new(),
         }
     }
@@ -69,7 +74,7 @@ impl ClassDiagram {
             Some(self.path.as_str())
         };
 
-        crate::mermaid_renderer::render_diagram(&self.diagram, title)
+        crate::mermaid_renderer::render_diagram(&self.diagram, title, self.direction)
     }
 
     pub fn add_class(